functions moved to /utils
